requirements.txt
streamlit~=1.31.0
langchain~=0.1.5
loguru~=0.7.2
langchain-openai~=0.0.5
langchain-experimental
langchain-community
duckduckgo-search
langchainhub
numexpr~=2.9.0
wikipedia~=1.4.0
python-dotenv
